feat(authenticate): update to new firebase-authentication driver
