+++ src/higher-order-components/authenticate/index.ts
@@ -2,29 +2,9 @@ import { Component as IComponent, Sources, Sinks } from '../../component/types';
 import { Stream } from 'most';
 import { merge } from 'ramda';
 import { model } from './model';
+import { AuthenticationMethod } from './types';
 
-export const GOOGLE: 'GOOGLE' = 'GOOGLE';
-export const FACEBOOK: 'FACEBOOK' = 'FACEBOOK';
-export { EMAIL_AND_PASSWORD } from '../../driver/firebase-authentication';
-
-export type GoogleAuthenticationMethod = {
-  method: 'GOOGLE';
-};
-
-export type FacebookAuthenticationMethod = {
-  method: 'FACEBOOK';
-}
-
-export type EmailAndPasswordAuthentincationMethod = {
-  method: 'EMAIL_AND_PASSWORD';
-  email: string;
-  password: string;
-}
-
-export type AuthenticationMethod =
-  GoogleAuthenticationMethod |
-  FacebookAuthenticationMethod |
-  EmailAndPasswordAuthentincationMethod;
+export * from './types';
 
 export type AuthenticationSinks = {
   authenticationMethod$: Stream<AuthenticationMethod>
